LRCI-7075 Rename build runners to match naming conventions
