+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/ArchiveBinariesCachePortalControllerBuildRunner.java
@@ -27,7 +27,7 @@ public class ArchiveBinariesCachePortalControllerBuildRunner
 	}
 
 	@Override
-	protected void invokeBuilds() {
+	protected void invokeBuild() {
 		StringBuilder sb = new StringBuilder();
 
 		String invocationJobURL = _getInvocationJobURL();
