+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/SingleUpstreamPortalControllerSingleSuiteBuildRunner.java
@@ -1,5 +1,5 @@
 /**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
@@ -12,20 +12,25 @@ import java.util.Map;
 import java.util.Properties;
 
 /**
- * @author Yi-Chen Tsai
+ * @author Michael Hashimoto
  */
-public class PortalTestSuiteUpstreamControllerSingleSuiteBuildRunner
+public class SingleUpstreamPortalControllerSingleSuiteBuildRunner
 	<S extends PortalTestSuiteUpstreamControllerBuildData>
-		extends PortalTestSuiteUpstreamControllerBuildRunner<S> {
+		extends BaseUpstreamPortalControllerBuildRunner<S> {
 
-	protected PortalTestSuiteUpstreamControllerSingleSuiteBuildRunner(
+	protected SingleUpstreamPortalControllerSingleSuiteBuildRunner(
 		S buildData) {
 
 		super(buildData);
 	}
 
 	@Override
-	protected void invokeBuilds() {
+	protected String getInvocationJobName() {
+		return "test-portal-upstream";
+	}
+
+	@Override
+	protected void invokeBuild() {
 		S buildData = getBuildData();
 
 		String testSuiteName = buildData.getTestSuiteName();
