+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/TestSuiteSingleUpstreamPortalControllerSingleSuiteBuildRunner.java
@@ -1,0 +1,30 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.jenkins.results.parser;
+
+/**
+ * @author Yi-Chen Tsai
+ */
+public class TestSuiteSingleUpstreamPortalControllerSingleSuiteBuildRunner
+	<S extends PortalTestSuiteUpstreamControllerBuildData>
+		extends SingleUpstreamPortalControllerSingleSuiteBuildRunner<S> {
+
+	protected TestSuiteSingleUpstreamPortalControllerSingleSuiteBuildRunner(
+		S buildData) {
+
+		super(buildData);
+	}
+
+	@Override
+	protected String getInvocationJobName() {
+		S buildData = getBuildData();
+
+		return JenkinsResultsParserUtil.combine(
+			"test-portal-testsuite-upstream(",
+			buildData.getPortalUpstreamBranchName(), ")");
+	}
+
+}
